Beethoven/requirements.txt
discord.py
PyNaCl
youtube-search-python==1.4.9
yt-dlp
youtube_transcript_api